ref(onboarding): Split aiohttp onboarding docs (#103037)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
