+++ static/app/gettingStartedDocs/python/aiohttp/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './aiohttp';
+import docs from '.';
 
 describe('aiohttp onboarding docs', () => {
   it('renders doc correctly', () => {
